Added json format and removed HTML pages as response
